Removed OAuthResponseController and moved all its API to UserController
