+++ Promact.Oauth.Server/src/Promact.Oauth.Server/Controllers/UserController.cs
@@ -10,6 +10,7 @@ using System.Collections.Generic;
 using Promact.Oauth.Server.Constants;
 using Promact.Oauth.Server.ExceptionHandler;
 using Promact.Oauth.Server.Services;
+using Exceptionless;
 
 namespace Promact.Oauth.Server.Controllers
 {
@@ -679,7 +680,202 @@ namespace Promact.Oauth.Server.Controllers
             List<UserAc> projectUsers = await _userRepository.GetProjectUsersByTeamLeaderIdAsync(teamLeaderId);
             return Ok(projectUsers);
         }
+        /**
+        * @api {get} api/users/user/{slackUserId}
+        * @apiVersion 1.0.0
+        * @apiName UserDetialBySlackUserIdAsync
+        * @apiGroup User
+        * @apiParam {string} Name  slackUserId
+        * @apiParamExample {json} Request-Example:
+        *        {
+        *             slackUserId : ADF4HY54H
+        *        }      
+        * @apiSuccessExample {json} Success-Response:
+        * HTTP/1.1 200 OK 
+        *   {
+        *       "Id":"34d1af3d-062f-4bcd-b6f9-b8fd5165e367",
+        *       "FirstName" : "ABC",
+        *       "Email" : "abc@promactinfo.com",
+        *       "LastName" : "DEF",
+        *       "SlackUserId" : "ADF4HY54H",
+        *   }
+        * @apiErrorExample {json} Error-Response:
+        * HTTP/1.1 400 Bad Request 
+        * {
+        *     "error" : "SlackUserNotFound"
+        * }
+        */
+        [Authorize(Policy = ReadUser)]
+        [HttpGet]
+        [Route("user/{slackUserId}")]
+        public async Task<IActionResult> UserDetialBySlackUserIdAsync(string slackUserId)
+        {
+            try
+            {
+                return Ok(await _userRepository.UserDetialByUserSlackIdAsync(slackUserId));
+            }
+            catch (SlackUserNotFound ex)
+            {
+                ex.ToExceptionless().Submit();
+                return BadRequest(ex.StackTrace);
+            }
+        }
 
+        /**
+        * @api {get} api/users/teamLeaders/{slackUserId}
+        * @apiVersion 1.0.0
+        * @apiName TeamLeaderByUsersSlackIdAsync
+        * @apiGroup User
+        * @apiParam {string} Name  slackUserId
+        * @apiParamExample {json} Request-Example:
+        *        {
+        *             slackUserId : ADF4HY54H
+        *        }      
+        * @apiSuccessExample {json} Success-Response:
+        * HTTP/1.1 200 OK 
+        *   {
+        *       [
+        *           {
+        *               "Email" : "abc@promactinfo.com",
+        *               "UserName" : "abc@promactinfo.com",
+        *               "SlackUserId" : "ADF4HY54H",
+        *           }
+        *       ]
+        *   }
+        * @apiErrorExample {json} Error-Response:
+        * HTTP/1.1 400 Bad Request 
+        * {
+        *     "error" : "SlackUserNotFound"
+        * }
+        */
+        [Authorize(Policy = ReadUser)]
+        [HttpGet]
+        [Route("teamLeaders/{slackUserId}")]
+        public async Task<IActionResult> TeamLeaderByUsersSlackIdAsync(string slackUserId)
+        {
+            try
+            {
+                return Ok(await _userRepository.TeamLeaderByUserSlackIdAsync(slackUserId));
+            }
+            catch (SlackUserNotFound ex)
+            {
+                ex.ToExceptionless().Submit();
+                return BadRequest(ex.StackTrace);
+            }
+        }
+
+        /**
+        * @api {get} api/users/managements
+        * @apiVersion 1.0.0
+        * @apiName ManagementDetailsAsync
+        * @apiGroup User    
+        * @apiSuccessExample {json} Success-Response:
+        * HTTP/1.1 200 OK 
+        *   {
+        *       [
+        *           {
+        *               "Email" : "abc@promactinfo.com",
+        *               "UserName" : "abc@promactinfo.com",
+        *               "SlackUserId" : "ADF4HY54H",
+        *           }
+        *       ]
+        *   }
+        * @apiErrorExample {json} Error-Response:
+        * HTTP/1.1 400 Bad Request 
+        * {
+        *     "error" : "FailedToFetchDataException"
+        * }
+        */
+        [Authorize(Policy = ReadUser)]
+        [HttpGet]
+        [Route("managements")]
+        public async Task<IActionResult> ManagementDetailsAsync()
+        {
+            try
+            {
+                return Ok(await _userRepository.ManagementDetailsAsync());
+            }
+            catch (FailedToFetchDataException ex)
+            {
+                ex.ToExceptionless().Submit();
+                return BadRequest(ex.StackTrace);
+            }
+        }
+
+        /**
+        * @api {get} api/users/leaveAllowed/{slackUserId}
+        * @apiVersion 1.0.0
+        * @apiName GetUserCasualLeaveBySlackIdAsync
+        * @apiGroup User    
+        * @apiParam {string} Name  slackUserId
+        * @apiParamExample {json} Request-Example:
+        *        {
+        *             slackUserId : ADF4HY54H
+        *        } 
+        * @apiSuccessExample {json} Success-Response:
+        * HTTP/1.1 200 OK 
+        *   {
+        *       "CasualLeave" : "14",
+        *       "SickLeave" : "7"
+        *   }
+        * @apiErrorExample {json} Error-Response:
+        * HTTP/1.1 400 Bad Request 
+        * {
+        *     "error" : "SlackUserNotFound"
+        * }
+        */
+        [Authorize(Policy = ReadUser)]
+        [HttpGet]
+        [Route("leaveAllowed/{slackUserId}")]
+        public async Task<IActionResult> GetUserCasualLeaveBySlackIdAsync(string slackUserId)
+        {
+            try
+            {
+                return Ok(await _userRepository.GetUserAllowedLeaveBySlackIdAsync(slackUserId));
+            }
+            catch (SlackUserNotFound ex)
+            {
+                ex.ToExceptionless().Submit();
+                return BadRequest(ex.StackTrace);
+            }
+        }
+
+        /**
+        * @api {get} api/users/userIsAdmin/{slackUserId}
+        * @apiVersion 1.0.0
+        * @apiName UserIsAdminAsync
+        * @apiGroup User    
+        * @apiParam {string} Name  slackUserId
+        * @apiParamExample {json} Request-Example:
+        *        {
+        *             slackUserId : ADF4HY54H
+        *        } 
+        * @apiSuccessExample {json} Success-Response:
+        * HTTP/1.1 200 OK 
+        *   {
+        *       "true"
+        *   }
+        * @apiErrorExample {json} Error-Response:
+        * HTTP/1.1 400 Bad Request 
+        * {
+        *     "error" : "SlackUserNotFound"
+        * }
+        */
+        [Authorize(Policy = ReadUser)]
+        [HttpGet]
+        [Route("userIsAdmin/{slackUserId}")]
+        public async Task<IActionResult> UserIsAdminAsync(string slackUserId)
+        {
+            try
+            {
+                return Ok(await _userRepository.IsAdminAsync(slackUserId));
+            }
+            catch (SlackUserNotFound ex)
+            {
+                ex.ToExceptionless().Submit();
+                return BadRequest(ex.StackTrace);
+            }
+        }
         #endregion
     }
 }
